added faker data for voters

+++ db/migrate/20170924185611_create_voters.rb
@@ -4,7 +4,7 @@ class CreateVoters < ActiveRecord::Migration[5.1]
       t.string :name # Name
       t.integer :age # Age
       t.text :location # Location
-      t.integer :phone_number # Phone number
+      t.bigint :phone_number # Phone number
       t.datetime :last_contact # Date last contacted
       t.timestamps
     end
